+++ tests/Bot/ProviderWrapperTest.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace seregazhuk\tests;
+namespace seregazhuk\tests\Bot;
 
 use PHPUnit_Framework_TestCase;
 use seregazhuk\PinterestBot\Api\Request;
+use seregazhuk\PinterestBot\Api\Response;
+use seregazhuk\PinterestBot\Helpers\Cookies;
 use seregazhuk\PinterestBot\Api\CurlHttpClient;
 use seregazhuk\PinterestBot\Api\Providers\Provider;
 use seregazhuk\PinterestBot\Api\Providers\ProviderWrapper;
-use seregazhuk\PinterestBot\Api\Response;
 
 class ProviderWrapperTest extends PHPUnit_Framework_TestCase
 {
@@ -44,7 +45,10 @@ class ProviderWrapperTest extends PHPUnit_Framework_TestCase
      */
     protected function createWrapper()
     {
-        $provider = new TestProvider(new Request(new CurlHttpClient()), new Response());
+        $request = new Request(new CurlHttpClient(new Cookies()));
+
+        $provider = new TestProvider($request, new Response());
+
         return new ProviderWrapper($provider);
     }
 }